Entryの削除機能の実装 #13

+++ public/js/req_method.js
@@ -1,9 +1,22 @@
 function del_like(id) {
-	 $.ajax({
-		  url: '/entry/'+id+'/like',
-		  type: 'DELETE',
-		  success: function(result) {
+	$.ajax({
+		url: '/entry/'+id+'/like',
+		type: 'DELETE',
+		success: function(result) {
+			location.href="/";
+		}
+	});
+}
+
+function del_entry(id) {
+	if(window.confirm('本当に削除してよろしいいですか？'));
+	{
+		$.ajax({
+			url: '/entry/'+id,
+			type: 'DELETE',
+			success: function(result) {
 				location.href="/";
-		  }
-	 });
+			}
+		});
+	}
 }
